Trying to decouple a little bit.
